backend/requirements.txt
# Core Django Framework
Django>=5.2,<6.0

# Base de Datos MySQL
mysqlclient>=2.2.0
PyMySQL>=1.1.0
cryptography>=41.0.0

# Configuración y Variables de Entorno
python-decouple>=3.8

# Generación de Diagramas DER
sqlalchemy>=2.0.0
graphviz>=0.20.0

# Reportes y Excel
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Gestión de Archivos
Pillow>=10.0.0

# API y Serialización
djangorestframework>=3.14.0
drf-spectacular>=0.27.0

# Utilidades
pytz>=2023.3
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-mock>=3.11.0
factory-boy>=3.3.0
faker>=19.0.0
coverage>=7.3.0
